+++ handler/post/send_post.go
@@ -7,7 +7,7 @@ import (
 )
 
 //参数检查要在api层做好，这里保证参数不为空
-func DealSendPost(userId int, title, content string) (err error) {
+func DealSendPost(userId int64, title, content string) (err error) {
 	defer func() {
 		if err != nil {
 			fmt.Println("handler层：DealSendPost err:", err)
@@ -17,7 +17,7 @@ func DealSendPost(userId int, title, content string) (err error) {
 		err = errors.New("您先前已经发布了一篇相同标题的帖子")
 		return
 	}
-	if err = mysql.AddNewPostRecord(int64(userId), title, content); err != nil {
+	if err = mysql.AddNewPostRecord(userId, title, content); err != nil {
 		err = errors.Wrap(err, "发布帖子失败，请稍后重试")
 		return
 	}
